+++ src/engine/kinematics/index.ts
@@ -1,3 +1,4 @@
 export * from './angleCalculator';
 export * from './emaFilter';
 export * from './parallaxCheck';
+export * from './poseDetector';