+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useCamera';
 export * from './useAudioCoach';
 export * from './useMedicalProfile';
+export * from './usePoseTracking';